+++ client/orm/internal/buffers/buffers.go
@@ -1,10 +1,10 @@
-// Copyright 2020
+// Copyright 2014 beego Author. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-// http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,24 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package orm
+package buffers
 
-import (
-	"reflect"
-	"testing"
+import "github.com/valyala/bytebufferpool"
 
-	"github.com/stretchr/testify/assert"
-)
+var _ Buffer = &bytebufferpool.ByteBuffer{}
 
-type NotApplicableModel struct {
-	Id int
+type Buffer interface {
+	Write(p []byte) (int, error)
+	WriteString(s string) (int, error)
+	WriteByte(c byte) error
 }
 
-func (n *NotApplicableModel) IsApplicableTableForDB(db string) bool {
-	return db == "default"
+func Get() Buffer {
+	return bytebufferpool.Get()
 }
 
-func TestIsApplicableTableForDB(t *testing.T) {
-	assert.False(t, isApplicableTableForDB(reflect.ValueOf(&NotApplicableModel{}), "defa"))
-	assert.True(t, isApplicableTableForDB(reflect.ValueOf(&NotApplicableModel{}), "default"))
+func Put(bf Buffer) {
+	bytebufferpool.Put(bf.(*bytebufferpool.ByteBuffer))
 }