+++ client/orm/internal/models/models_utils_test.go
@@ -1,10 +1,10 @@
-// Copyright 2014 beego Author. All Rights Reserved.
+// Copyright 2020
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//      http://www.apache.org/licenses/LICENSE-2.0
+// http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,27 +12,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package orm
+package models
 
 import (
+	"reflect"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
-func TestCamelString(t *testing.T) {
-	snake := []string{"pic_url", "hello_world_", "hello__World", "_HelLO_Word", "pic_url_1", "pic_url__1"}
-	camel := []string{"PicUrl", "HelloWorld", "HelloWorld", "HelLOWord", "PicUrl1", "PicUrl1"}
+type NotApplicableModel struct {
+	Id int
+}
 
-	answer := make(map[string]string)
-	for i, v := range snake {
-		answer[v] = camel[i]
-	}
+func (n *NotApplicableModel) IsApplicableTableForDB(db string) bool {
+	return db == "default"
+}
 
-	for _, v := range snake {
-		res := camelString(v)
-		if res != answer[v] {
-			t.Error("Unit Test Fail:", v, res, answer[v])
-		}
-	}
+func TestIsApplicableTableForDB(t *testing.T) {
+	assert.False(t, IsApplicableTableForDB(reflect.ValueOf(&NotApplicableModel{}), "defa"))
+	assert.True(t, IsApplicableTableForDB(reflect.ValueOf(&NotApplicableModel{}), "default"))
 }
 
 func TestSnakeString(t *testing.T) {
@@ -45,7 +44,7 @@ func TestSnakeString(t *testing.T) {
 	}
 
 	for _, v := range camel {
-		res := snakeString(v)
+		res := SnakeString(v)
 		if res != answer[v] {
 			t.Error("Unit Test Fail:", v, res, answer[v])
 		}
@@ -62,7 +61,24 @@ func TestSnakeStringWithAcronym(t *testing.T) {
 	}
 
 	for _, v := range camel {
-		res := snakeStringWithAcronym(v)
+		res := SnakeStringWithAcronym(v)
+		if res != answer[v] {
+			t.Error("Unit Test Fail:", v, res, answer[v])
+		}
+	}
+}
+
+func TestCamelString(t *testing.T) {
+	snake := []string{"pic_url", "hello_world_", "hello__World", "_HelLO_Word", "pic_url_1", "pic_url__1"}
+	camel := []string{"PicUrl", "HelloWorld", "HelloWorld", "HelLOWord", "PicUrl1", "PicUrl1"}
+
+	answer := make(map[string]string)
+	for i, v := range snake {
+		answer[v] = camel[i]
+	}
+
+	for _, v := range snake {
+		res := CamelString(v)
 		if res != answer[v] {
 			t.Error("Unit Test Fail:", v, res, answer[v])
 		}