+++ client/orm/orm_queryset.go
@@ -18,6 +18,10 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/beego/beego/v2/client/orm/internal/utils"
+
+	"github.com/beego/beego/v2/client/orm/internal/models"
+
 	"github.com/beego/beego/v2/client/orm/clauses/order_clause"
 	"github.com/beego/beego/v2/client/orm/hints"
 )
@@ -54,7 +58,7 @@ func ColValue(opt operator, value interface{}) interface{} {
 	default:
 		panic(fmt.Errorf("orm.ColValue wrong operator"))
 	}
-	v, err := StrTo(ToStr(value)).Int64()
+	v, err := utils.StrTo(utils.ToStr(value)).Int64()
 	if err != nil {
 		panic(fmt.Errorf("orm.ColValue doesn't support non string/numeric type, %s", err))
 	}
@@ -66,7 +70,7 @@ func ColValue(opt operator, value interface{}) interface{} {
 
 // real query struct
 type querySet struct {
-	mi        *modelInfo
+	mi        *models.ModelInfo
 	cond      *Condition
 	related   []string
 	relDepth  int
@@ -113,13 +117,13 @@ func (o querySet) Exclude(expr string, args ...interface{}) QuerySeter {
 
 // set offset number
 func (o *querySet) setOffset(num interface{}) {
-	o.offset = ToInt64(num)
+	o.offset = utils.ToInt64(num)
 }
 
 // add LIMIT value.
 // args[0] means offset, e.g. LIMIT num,offset.
 func (o querySet) Limit(limit interface{}, args ...interface{}) QuerySeter {
-	o.limit = ToInt64(limit)
+	o.limit = utils.ToInt64(limit)
 	if len(args) > 0 {
 		o.setOffset(args[0])
 	}
@@ -273,7 +277,7 @@ func (o *querySet) PrepareInsertWithCtx(ctx context.Context) (Inserter, error) {
 }
 
 // query all data and map to containers.
-// cols means the columns when querying.
+// cols means the Columns when querying.
 func (o *querySet) All(container interface{}, cols ...string) (int64, error) {
 	return o.AllWithCtx(context.Background(), container, cols...)
 }
@@ -283,7 +287,7 @@ func (o *querySet) AllWithCtx(ctx context.Context, container interface{}, cols .
 }
 
 // query one row data and map to containers.
-// cols means the columns when querying.
+// cols means the Columns when querying.
 func (o *querySet) One(container interface{}, cols ...string) error {
 	return o.OneWithCtx(context.Background(), container, cols...)
 }
@@ -366,7 +370,7 @@ func (o *querySet) RowsToStruct(ptrStruct interface{}, keyCol, valueCol string) 
 }
 
 // create new QuerySeter.
-func newQuerySet(orm *ormBase, mi *modelInfo) QuerySeter {
+func newQuerySet(orm *ormBase, mi *models.ModelInfo) QuerySeter {
 	o := new(querySet)
 	o.mi = mi
 	o.orm = orm