+++ client/orm/orm_raw.go
@@ -20,6 +20,10 @@ import (
 	"reflect"
 	"time"
 
+	"github.com/beego/beego/v2/client/orm/internal/utils"
+
+	"github.com/beego/beego/v2/client/orm/internal/models"
+
 	"github.com/pkg/errors"
 )
 
@@ -95,7 +99,7 @@ func (o *rawSet) setFieldValue(ind reflect.Value, value interface{}) {
 		} else if v, ok := value.(bool); ok {
 			ind.SetBool(v)
 		} else {
-			v, _ := StrTo(ToStr(value)).Bool()
+			v, _ := utils.StrTo(utils.ToStr(value)).Bool()
 			ind.SetBool(v)
 		}
 
@@ -103,7 +107,7 @@ func (o *rawSet) setFieldValue(ind reflect.Value, value interface{}) {
 		if value == nil {
 			ind.SetString("")
 		} else {
-			ind.SetString(ToStr(value))
+			ind.SetString(utils.ToStr(value))
 		}
 
 	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
@@ -117,7 +121,7 @@ func (o *rawSet) setFieldValue(ind reflect.Value, value interface{}) {
 			case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
 				ind.SetInt(int64(val.Uint()))
 			default:
-				v, _ := StrTo(ToStr(value)).Int64()
+				v, _ := utils.StrTo(utils.ToStr(value)).Int64()
 				ind.SetInt(v)
 			}
 		}
@@ -132,7 +136,7 @@ func (o *rawSet) setFieldValue(ind reflect.Value, value interface{}) {
 			case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
 				ind.SetUint(val.Uint())
 			default:
-				v, _ := StrTo(ToStr(value)).Uint64()
+				v, _ := utils.StrTo(utils.ToStr(value)).Uint64()
 				ind.SetUint(v)
 			}
 		}
@@ -145,7 +149,7 @@ func (o *rawSet) setFieldValue(ind reflect.Value, value interface{}) {
 			case reflect.Float64:
 				ind.SetFloat(val.Float())
 			default:
-				v, _ := StrTo(ToStr(value)).Float64()
+				v, _ := utils.StrTo(utils.ToStr(value)).Float64()
 				ind.SetFloat(v)
 			}
 		}
@@ -170,20 +174,20 @@ func (o *rawSet) setFieldValue(ind reflect.Value, value interface{}) {
 			if str != "" {
 				if len(str) >= 19 {
 					str = str[:19]
-					t, err := time.ParseInLocation(formatDateTime, str, o.orm.alias.TZ)
+					t, err := time.ParseInLocation(utils.FormatDateTime, str, o.orm.alias.TZ)
 					if err == nil {
 						t = t.In(DefaultTimeLoc)
 						ind.Set(reflect.ValueOf(t))
 					}
 				} else if len(str) >= 10 {
 					str = str[:10]
-					t, err := time.ParseInLocation(formatDate, str, DefaultTimeLoc)
+					t, err := time.ParseInLocation(utils.FormatDate, str, DefaultTimeLoc)
 					if err == nil {
 						ind.Set(reflect.ValueOf(t))
 					}
 				} else if len(str) >= 8 {
 					str = str[:8]
-					t, err := time.ParseInLocation(formatTime, str, DefaultTimeLoc)
+					t, err := time.ParseInLocation(utils.FormatTime, str, DefaultTimeLoc)
 					if err == nil {
 						ind.Set(reflect.ValueOf(t))
 					}
@@ -287,7 +291,7 @@ func (o *rawSet) QueryRow(containers ...interface{}) error {
 		refs  = make([]interface{}, 0, len(containers))
 		sInds []reflect.Value
 		eTyps []reflect.Type
-		sMi   *modelInfo
+		sMi   *models.ModelInfo
 	)
 	structMode := false
 	for _, container := range containers {
@@ -313,7 +317,7 @@ func (o *rawSet) QueryRow(containers ...interface{}) error {
 			}
 
 			structMode = true
-			fn := getFullName(typ)
+			fn := models.GetFullName(typ)
 			if mi, ok := defaultModelCache.getByFullName(fn); ok {
 				sMi = mi
 			}
@@ -370,16 +374,16 @@ func (o *rawSet) QueryRow(containers ...interface{}) error {
 
 			if sMi != nil {
 				for _, col := range columns {
-					if fi := sMi.fields.GetByColumn(col); fi != nil {
+					if fi := sMi.Fields.GetByColumn(col); fi != nil {
 						value := reflect.ValueOf(columnsMp[col]).Elem().Interface()
-						field := ind.FieldByIndex(fi.fieldIndex)
-						if fi.fieldType&IsRelField > 0 {
-							mf := reflect.New(fi.relModelInfo.addrField.Elem().Type())
+						field := ind.FieldByIndex(fi.FieldIndex)
+						if fi.FieldType&IsRelField > 0 {
+							mf := reflect.New(fi.RelModelInfo.AddrField.Elem().Type())
 							field.Set(mf)
-							field = mf.Elem().FieldByIndex(fi.relModelInfo.fields.pk.fieldIndex)
+							field = mf.Elem().FieldByIndex(fi.RelModelInfo.Fields.Pk.FieldIndex)
 						}
-						if fi.isFielder {
-							fd := field.Addr().Interface().(Fielder)
+						if fi.IsFielder {
+							fd := field.Addr().Interface().(models.Fielder)
 							err := fd.SetRaw(value)
 							if err != nil {
 								return errors.Errorf("set raw error:%s", err)
@@ -406,12 +410,12 @@ func (o *rawSet) QueryRow(containers ...interface{}) error {
 						// thanks @Gazeboxu.
 						tags := structTagMap[fe.Tag]
 						if tags == nil {
-							_, tags = parseStructTag(fe.Tag.Get(defaultStructTagName))
+							_, tags = models.ParseStructTag(fe.Tag.Get(models.DefaultStructTagName))
 							structTagMap[fe.Tag] = tags
 						}
 						var col string
 						if col = tags["column"]; col == "" {
-							col = nameStrategyMap[nameStrategy](fe.Name)
+							col = models.NameStrategyMap[models.NameStrategy](fe.Name)
 						}
 						if v, ok := columnsMp[col]; ok {
 							value := reflect.ValueOf(v).Elem().Interface()
@@ -449,7 +453,7 @@ func (o *rawSet) QueryRows(containers ...interface{}) (int64, error) {
 		refs  = make([]interface{}, 0, len(containers))
 		sInds []reflect.Value
 		eTyps []reflect.Type
-		sMi   *modelInfo
+		sMi   *models.ModelInfo
 	)
 	structMode := false
 	for _, container := range containers {
@@ -474,7 +478,7 @@ func (o *rawSet) QueryRows(containers ...interface{}) (int64, error) {
 			}
 
 			structMode = true
-			fn := getFullName(typ)
+			fn := models.GetFullName(typ)
 			if mi, ok := defaultModelCache.getByFullName(fn); ok {
 				sMi = mi
 			}
@@ -537,16 +541,16 @@ func (o *rawSet) QueryRows(containers ...interface{}) (int64, error) {
 
 			if sMi != nil {
 				for _, col := range columns {
-					if fi := sMi.fields.GetByColumn(col); fi != nil {
+					if fi := sMi.Fields.GetByColumn(col); fi != nil {
 						value := reflect.ValueOf(columnsMp[col]).Elem().Interface()
-						field := ind.FieldByIndex(fi.fieldIndex)
-						if fi.fieldType&IsRelField > 0 {
-							mf := reflect.New(fi.relModelInfo.addrField.Elem().Type())
+						field := ind.FieldByIndex(fi.FieldIndex)
+						if fi.FieldType&IsRelField > 0 {
+							mf := reflect.New(fi.RelModelInfo.AddrField.Elem().Type())
 							field.Set(mf)
-							field = mf.Elem().FieldByIndex(fi.relModelInfo.fields.pk.fieldIndex)
+							field = mf.Elem().FieldByIndex(fi.RelModelInfo.Fields.Pk.FieldIndex)
 						}
-						if fi.isFielder {
-							fd := field.Addr().Interface().(Fielder)
+						if fi.IsFielder {
+							fd := field.Addr().Interface().(models.Fielder)
 							err := fd.SetRaw(value)
 							if err != nil {
 								return 0, errors.Errorf("set raw error:%s", err)
@@ -570,10 +574,10 @@ func (o *rawSet) QueryRows(containers ...interface{}) (int64, error) {
 							recursiveSetField(f)
 						}
 
-						_, tags := parseStructTag(fe.Tag.Get(defaultStructTagName))
+						_, tags := models.ParseStructTag(fe.Tag.Get(models.DefaultStructTagName))
 						var col string
 						if col = tags["column"]; col == "" {
-							col = nameStrategyMap[nameStrategy](fe.Name)
+							col = models.NameStrategyMap[models.NameStrategy](fe.Name)
 						}
 						if v, ok := columnsMp[col]; ok {
 							value := reflect.ValueOf(v).Elem().Interface()
@@ -837,7 +841,7 @@ func (o *rawSet) queryRowsTo(container interface{}, keyCol, valueCol string) (in
 			}
 
 		default:
-			if id := ind.FieldByName(camelString(key)); id.IsValid() {
+			if id := ind.FieldByName(models.CamelString(key)); id.IsValid() {
 				o.setFieldValue(id, reflect.ValueOf(refs[valueIndex]).Elem().Interface())
 			}
 		}